nit: Formatting, refactor
