+++ webapp/config.defaults.json
@@ -13,7 +13,7 @@
     "DISABLE_BACKGROUND_WORKER": false,
     "TASK_BASE_URL": "http://sovietov.com/kispython",
     "FINAL_TASKS": {
-        "1": [0, 1, 2, 3, 4],
-        "2": [5, 6, 7, 8]
+        "0": [0, 1, 2, 3, 4],
+        "1": [5, 6, 7, 8]
     }
 }
