+++ webapp/views.py
@@ -1,4 +1,3 @@
-import functools
 from typing import Union
 
 from werkzeug.exceptions import HTTPException
@@ -10,18 +9,21 @@ from flask import make_response, redirect, render_template, request, url_for
 from webapp.forms import MessageForm
 from webapp.managers import AppConfigManager, ExportManager, StatusManager
 from webapp.repositories import AppDatabase
-from webapp.utils import get_exception_info, get_real_ip
+from webapp.utils import get_exception_info, get_real_ip, require_token
 
 
 blueprint = Blueprint("views", __name__)
 config = AppConfigManager(lambda: app.config)
 db = AppDatabase(lambda: config.config.connection_string)
+
 exports = ExportManager(
-    db.groups,
-    db.messages,
-    db.statuses,
-    db.variants,
-    db.tasks)
+    groups=db.groups,
+    messages=db.messages,
+    statuses=db.statuses,
+    variants=db.variants,
+    tasks=db.tasks
+)
+
 statuses = StatusManager(
     tasks=db.tasks,
     groups=db.groups,
@@ -32,28 +34,9 @@ statuses = StatusManager(
 )
 
 
-def require_final_token():
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            if len(kwargs) == 0 or config.config.final_token \
-                                   != kwargs['token']:
-                return render_template(
-                       "error.jinja",
-                       error_code=401,
-                       error_message="Unauthorized",
-                       error_redirect="/",
-                )
-            return func(*args, **kwargs)
-        return wrapper
-    return decorator
-
-
 @blueprint.route("/", methods=["GET"])
 def dashboard():
     groupings = db.groups.get_groupings()
-    if config.config.final_tasks:
-        return render_template("examstatus.jinja")
     return render_template("dashboard.jinja", groupings=groupings)
 
 
@@ -107,21 +90,21 @@ def export(s: str, token: str, count: Union[int, None]):
 
 
 @blueprint.route("/exams/<token>", methods=["GET"])
-@require_final_token()
+@require_token(lambda: config.config.final_token)
 def exams(token: str):
     groups = db.groups.get_all()
     return render_template("exams.jinja", groups=groups, token=token)
 
 
 @blueprint.route("/exams/<token>", methods=["POST"])
-@require_final_token()
+@require_token(lambda: config.config.final_token)
 def enter_exam_group(token: str):
     group_id = request.form.get('group')
     return redirect(url_for("views.pre_exam", group_id=group_id, token=token))
 
 
 @blueprint.route("/exams/<token>/<group_id>", methods=["GET"])
-@require_final_token()
+@require_token(lambda: config.config.final_token)
 def pre_exam(group_id: int, token: str):
     group = db.groups.get_by_id(group_id)
     seed = db.seeds.get_final_seed(group_id)
@@ -129,7 +112,7 @@ def pre_exam(group_id: int, token: str):
 
 
 @blueprint.route("/exams/<token>/<group_id>", methods=["POST"])
-@require_final_token()
+@require_token(lambda: config.config.final_token)
 def exam(group_id: int, token: str):
     seed = db.seeds.get_final_seed(group_id)
     if seed is None:
@@ -142,7 +125,7 @@ def exam(group_id: int, token: str):
 
 
 @blueprint.route("/exams/<token>/<group_id>/score_csv", methods=["POST"])
-@require_final_token()
+@require_token(lambda: config.config.final_token)
 def score_csv(group_id: int, token: str):
     delimiter = request.form.get('delimiter')
     value = exports.export_exam_results(group_id, delimiter)
